pyproject.toml
[project]
name = "iecdt_lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "notebook>=7.3.2",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "wandb==0.24.0",
    "zarr>=3.0.1",
]

[tool.setuptools]
packages = ["iecdt_lab"]
